disclaimer for limited v12 support
At the current moment I only have the resources to develop for Foundry v11, with plans to move to v12 at some point. 

Provided there is not signifigant code changes that need to be made, this module is set to work with v12 already, but there may be some unforseen issues present.