updated severity logic
The original severity logic was basically ~60% minor, ~30% Heavy, ~10% Massive. The new logic makes it so that the severity is now linked to travel roll die size (but not the roll itself).

randomSeverity now has the difficulty, or the travel roll die size, passed in. And it should now "roll" that dice, seperate from the travel roll itself, to determine the severity.